1_first_demo/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3
msal==1.30.0
PyJWT==2.9.0
cryptography==43.0.1
